Filter_Classes_&_Delete_CheckBox

+++ app/Http/Controllers/Grades/GradeController.php
@@ -3,6 +3,7 @@
 
 namespace App\Http\Controllers\Grades;
 use App\Http\Controllers\Controller;
+use App\Models\Classroom;
 use CodeZero\UniqueTranslation\UniqueTranslationRule;
 
 use App\Http\Requests\StoreGrades;
@@ -90,10 +91,22 @@ class GradeController extends Controller
    */
   public function destroy(Request $request)
   {
+      $MyClass_id = Classroom::where('Grade_id',$request->id)->pluck('Grade_id');
 
-    $Grades = Grade::findOrFail($request->id)->delete();
-    toastr()->error(trans('messages.Delete'));
-    return redirect()->route('Grades.index');
+      if($MyClass_id->count() == 0){
+
+          $Grades = Grade::findOrFail($request->id)->delete();
+          toastr()->error(trans('messages.Delete'));
+          return redirect()->route('Grades.index');
+      }
+
+      else{
+
+          toastr()->error(trans('Grades_trans.delete_Grade_Error'));
+          return redirect()->route('Grades.index');
+
+      }
+
 
   }
